+++ foundrytools_cli_2/cli/converter/cli.py
@@ -6,8 +6,7 @@ from typing import Any, Literal, Optional
 import click
 from foundrytools import Font
 
-from foundrytools_cli_2.cli.base_command import BaseCommand
-from foundrytools_cli_2.cli.shared_callbacks import choice_to_int_callback
+from foundrytools_cli_2.cli import BaseCommand, choice_to_int_callback
 from foundrytools_cli_2.cli.task_runner import TaskRunner
 
 cli = click.Group("converter", help="Font conversion utilities.")
