+++ foundrytools_cli_2/cli/font/cli.py
@@ -4,7 +4,7 @@ from typing import Any
 import click
 from foundrytools.core.font import Font
 
-from foundrytools_cli_2.cli.base_command import BaseCommand
+from foundrytools_cli_2.cli import BaseCommand
 from foundrytools_cli_2.cli.logger import logger
 from foundrytools_cli_2.cli.task_runner import TaskRunner
