+++ foundrytools_cli_2/cli/gsub/cli.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import click
 from foundrytools import Font
 
-from foundrytools_cli_2.cli.base_command import BaseCommand
+from foundrytools_cli_2.cli import BaseCommand
 from foundrytools_cli_2.cli.logger import logger
 from foundrytools_cli_2.cli.task_runner import TaskRunner
