+++ foundrytools_cli_2/cli/post/cli.py
@@ -4,9 +4,8 @@ from pathlib import Path
 import click
 from foundrytools import Font
 
-from foundrytools_cli_2.cli.base_command import BaseCommand
+from foundrytools_cli_2.cli import BaseCommand, ensure_at_least_one_param
 from foundrytools_cli_2.cli.logger import logger
-from foundrytools_cli_2.cli.shared_callbacks import ensure_at_least_one_param
 from foundrytools_cli_2.cli.task_runner import TaskRunner
