+++ foundrytools_cli_2/cli/utils/cli.py
@@ -1,21 +1,44 @@
 # pylint: disable=import-outside-toplevel
-import typing as t
 from pathlib import Path
+from typing import Any, Callable, cast
 
 import click
 
-from foundrytools_cli_2.cli.shared_options import input_path_argument, recursive_flag
+from foundrytools_cli_2.cli import make_options
 from foundrytools_cli_2.cli.task_runner import TaskRunner
 from foundrytools_cli_2.cli.utils.options import font_organizer_options, rename_source_option
+
+
+def recursive_flag() -> Callable:
+    """
+    Add the ``recursive`` option to a click command.
+
+    :return: A decorator that adds the ``recursive`` option to a click command
+    :rtype: Callable
+    """
+    _recursive_flag = [
+        click.option(
+            "-r",
+            "--recursive",
+            is_flag=True,
+            default=False,
+            help="""
+            If ``input_path`` is a directory, the font finder will search for fonts recursively in
+            subdirectories.
+            """,
+        )
+    ]
+    return make_options(_recursive_flag)
+
 
 cli = click.Group(help="Miscellaneous utilities.")
 
 
 @cli.command("font-renamer")
-@input_path_argument()
+@click.argument("input_path", type=click.Path(exists=True, resolve_path=True, path_type=Path))
 @rename_source_option()
 @recursive_flag()
-def font_renamer(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
+def font_renamer(input_path: Path, **options: dict[str, Any]) -> None:
     """
     Renames the given font files.
     """
@@ -27,15 +50,15 @@ def font_renamer(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
 
 
 @cli.command("font-organizer")
-@input_path_argument()
+@click.argument("input_path", type=click.Path(exists=True, resolve_path=True, path_type=Path))
 @font_organizer_options()
 @recursive_flag()
-def font_organizer(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
+def font_organizer(input_path: Path, **options: dict[str, Any]) -> None:
     """
     Organizes the given font files based on specified sorting options.
     """
     # This is a workaround to make the task work with the current TaskRunner
-    options["in_path"] = t.cast(t.Any, input_path)
+    options["in_path"] = cast(Any, input_path)
 
     from foundrytools_cli_2.cli.utils.tasks.font_organizer import main as task
 
@@ -45,7 +68,7 @@ def font_organizer(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
 
 
 @cli.command("sync-timestamps")
-@input_path_argument()
+@click.argument("input_path", type=click.Path(exists=True, resolve_path=True, path_type=Path))
 @recursive_flag()
 def align_timestamps(input_path: Path, recursive: bool = False) -> None:
     """
